[Fix] #16 - 찌르기 UI 수정 및 error handling
